Renamed package structured and dropdowns completed
