+++ src/main/java/com/selenium/dropdowns/Main.java
@@ -1,4 +1,4 @@
-package com.selenium;
+package com.selenium.dropdowns;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
